+++ types/global.d.ts
@@ -8,5 +8,7 @@ declare global
 
     type playerStateName = 'running' | 'airborne' | 'groundCling' | 'cling' | 'punch' | 'dive' | 'uppercut'
 
+    type GunDirection = undefined | 'up' | 'upward' | 'forward' | 'downward' | 'down';
+
     var __DEV__: boolean;
 }